restaurant-5 Completed upto section11

+++ Restaurants.Application/Dishes/Commands/DeleteDish/DeleteDishCommandHandler.cs
@@ -1,12 +1,14 @@
 ﻿using MediatR;
 using Microsoft.Extensions.Logging;
+using Restaurants.Domain.Constants;
 using Restaurants.Domain.Entities;
 using Restaurants.Domain.Exceptions;
+using Restaurants.Domain.Interfaces;
 using Restaurants.Domain.Repositories;
 
 namespace Restaurants.Application.Dishes.Commands.DeleteDish
 {
-    public class DeleteDishCommandHandler(ILogger<DeleteDishCommandHandler> _logger, IRestaurantsRepository restaurantsRepository, IDishesRepository dishesRepository) : IRequestHandler<DeleteDishCommand>
+    public class DeleteDishCommandHandler(ILogger<DeleteDishCommandHandler> _logger, IRestaurantsRepository restaurantsRepository, IDishesRepository dishesRepository, IRestaurantAuthorizationService restaurantAuthorizationService) : IRequestHandler<DeleteDishCommand>
     {
         public async Task Handle(DeleteDishCommand request, CancellationToken cancellationToken)
         {
@@ -14,6 +16,9 @@ namespace Restaurants.Application.Dishes.Commands.DeleteDish
 
             var restaurant = await restaurantsRepository.GetByIdAsync(request.RestaurantId);
             if (restaurant == null) throw new NotFoundException(nameof(Restaurant), request.RestaurantId.ToString());
+
+            if (!restaurantAuthorizationService.Authorize(restaurant, ResourceOperation.Delete))
+                throw new ForbidException();
 
             await dishesRepository.Delete(restaurant.Dishes);
         }
